Add RectFixedGenerator
